+++ test/functional/Loader/AbstractLoopMachineModuleLoaderTest.php
