+++ test/functional/Loader/AbstractModuleLoaderTest.php
