+++ test/unit/Loader/LoopMachineModuleLoaderTest.php
